+++ anne_auth/app/controllers/anne_auth/admin/sessions_controller.rb
@@ -1,6 +1,8 @@
 module AnneAuth
   module Admin
     class SessionsController < AnneAuth::ApplicationController
+      include AdminAuthentication
+
       layout "admin_auth"
 
       rate_limit to: 10, within: 3.minutes, only: :create, with: -> { redirect_to auth_route(:admin_login_path), alert: "時間をおいて再度お試しください。" }
@@ -11,7 +13,7 @@ module AnneAuth
 
       def create
         if admin_user = AnneAuth.configuration.admin_user_class.authenticate_by(session_params)
-          admin_user.update!(last_sign_in_at: Time.current)
+          admin_user.update!(last_sign_in_at: Time.current) if admin_user.respond_to?(:last_sign_in_at)
           start_new_admin_session_for(admin_user)
           redirect_to after_admin_authentication_url, notice: "ログインしました。"
         else