+++ anne_auth/test/models/anne_auth/model_namespace_test.rb
@@ -10,12 +10,16 @@ class AnneAuth::ModelNamespaceTest < ActiveSupport::TestCase
     assert_operator CustomerAccountPasswordResetToken, :<, AnneAuth::AccountPasswordResetToken
   end
 
-  test "admin sessions default to the engine model without a host Session wrapper" do
+  test "admin sessions default to the legacy engine model without a host Session wrapper" do
     assert_not Object.const_defined?(:Session, false)
+    assert_equal "AdminUser", AnneAuth.configuration.admin_user_class_name
+    assert_equal AdminUser, AnneAuth.configuration.admin_user_class
     assert_equal "AnneAuth::AdminSession", AnneAuth.configuration.admin_session_class_name
     assert_equal AnneAuth::AdminSession, AnneAuth.configuration.admin_session_class
+    assert_equal :admin_user_id, AnneAuth.configuration.admin_session_user_foreign_key
     assert_equal "AnneAuth::AdminSession", AnneAuth::AdminUser.reflect_on_association(:sessions).class_name
     assert_equal AnneAuth.configuration.admin_user_class, AnneAuth::AdminSession.reflect_on_association(:admin_user).klass
+    assert_equal :admin_user_id, AnneAuth::AdminSession.reflect_on_association(:admin_user).foreign_key.to_sym
   end
 
   test "engine account does not include host project extensions" do