+++ examples/customer_management/config/initializers/anne_auth.rb
@@ -1,6 +1,6 @@
 AnneAuth.configure do |config|
   config.mailer_from = "noreply@example.com"
   config.google_oauth_enabled = false
-
-  config.after_admin_login_path = ->(controller, _admin_user) { controller.main_app.admin_root_path }
+  config.account_class_name = "Account"
+  config.after_account_login_path = ->(controller, _account) { controller.main_app.admin_root_path }
 end